Fixed CodeFactor issue
