+++ src/NHibernate.Test/NHSpecificTest/GH2029/Fixture.cs
@@ -7,15 +7,6 @@ using NUnit.Framework;
 
 namespace NHibernate.Test.NHSpecificTest.GH2029
 {
-	public class TestClass
-	{
-		public virtual int Id { get; set; }
-		public virtual int? NullableInt32Prop { get; set; }
-		public virtual int Int32Prop { get; set; }
-		public virtual long? NullableInt64Prop { get; set; }
-		public virtual long Int64Prop { get; set; }
-	}
-
 	[TestFixture]
 	public class Fixture : TestCaseMappingByCode
 	{
